requirements.txt
streamlit==1.31.0
pandas==2.1.4
plotly==5.18.0
azure-storage-blob==12.19.0
python-dotenv==1.0.0